+++ db/queries/tenants.sql
@@ -6,3 +6,6 @@ SELECT * FROM tenants WHERE id = $1;
 
 -- name: ResolveTenant :one
 SELECT * FROM resolve_tenant(@slug);
+
+-- name: ListUserTenants :many
+SELECT * FROM user_tenants WHERE user_id = @user_id ORDER BY name;